Implemented search functions and created utils module
