+++ main.py
@@ -1,8 +1,4 @@
-from bs4 import BeautifulSoup
-from tqdm import tqdm
-import requests
-import os
-import pickle
+import xkcd_utils
 
 
 class xkcdcomic():
@@ -16,59 +12,6 @@ class xkcdcomic():
 
     def __str__(self):
         return(("{}: {}").format(self.title, self.link))
-
-
-def load_database():
-    database = list()
-    if(os.path.isfile('xkcd_database.pickle')):
-        with open('xkcd_database.pickle', 'rb') as f:
-            database = pickle.load(f)
-
-    # get number of newest comic
-    newest_num = requests.get("http://xkcd.com/info.0.json").json()['num']
-    if(len(database) + 1 == newest_num):
-        print("Database up to date!")
-        return database
-    else:
-        print("Generating database...")
-        for i in tqdm(range(len(database), newest_num + 1)):
-            # create database of comics to search through
-            num = str(i)
-            url = ("http://xkcd.com/{}/info.0.json").format(num)
-            try:
-                jsondata = requests.get(url).json()
-                comic = xkcdcomic(jsondata)
-                entry = (comic.number, comic.title, comic)
-                database.append(entry)
-            except:
-                continue
-        with open('xkcd_database.pickle', 'wb') as f:
-            # save database to file
-            pickle.dump(database, f)
-        return database
-
-
-def download_comic(img):
-    with open("img.png", 'wb') as pic:
-        for chunk in img.iter_content(4096):
-            pic.write(chunk)
-
-
-def show_prompt(comic):
-    print(("Comic title: {}").format(comic.title))
-    print(("Alt text: {}").format(comic.alt))
-    print("Opening comic...")
-    os.system("open img.png")
-    input("Press ENTER to delete the comic from storage")
-    os.system("rm -rf img.png")
-
-
-def download_and_show(comic):
-    url = comic.link
-    img = requests.get(url)
-    download_comic(img)
-    os.system("open img.png")
-    show_prompt(comic)
 
 
 def num_find(database):
@@ -100,6 +43,29 @@ def title_find(database):
     download_and_show(comic)
 
 
+def title_search(database):
+    keyword = input("Enter search term: ")
+
+    possible_comics = list()
+    for entry in database:
+        if keyword.lower() in entry[1].lower():
+            possible_comics.append(entry)
+
+    num_comics = len(possible_comics)
+    if num_comics == 0:
+        print("Couldn't find any comics with that search term!")
+        return
+
+    print("Results:")
+    for i in range(len(possible_comics)):
+        print(("{}) {}").format(i + 1, possible_comics[i][2].title))
+
+    # TODO: Check if integer
+    choice = int(input("Enter the number of the comic you want to see: "))
+    comic = possible_comics[choice - 1][2]
+    xkcd_utils.download_and_show(comic)
+
+
 def show_menu(database):
     print("What do you want to do? ")
     print("      To find a comic by number, type 'n'")
@@ -112,13 +78,13 @@ def show_menu(database):
     elif choice == "t":
         title_find(database)
     elif choice == "s":
-        pass
+        title_search(database)
     else:
         print("Invalid choice!")
 
 
 def main():
-    database = load_database()
+    database = xkcd_utils.load_database()
     show_menu(database)
